修正了在初始化glog库时的线程安全
为了方便调试，引入了glog工程，在D:\SDK文件夹

+++ dlog/Common/MemoryLog.cpp
@@ -36,7 +36,7 @@ namespace dxlib
         return _memLogQueue.try_dequeue(msg);
     }
 
-    int MemoryLog::count()
+    size_t MemoryLog::count()
     {
         return _memLogQueue.size_approx();
     }
